+++ mayan/apps/document_states/tests/test_workflow_actions.py
@@ -7,7 +7,7 @@ from mayan.apps.testing.tests.mixins import TestServerTestCaseMixin
 from mayan.apps.testing.tests.mocks import request_method_factory
 
 from ..literals import WORKFLOW_ACTION_ON_ENTRY
-from ..models import Workflow
+from ..models import Workflow, WorkflowInstance
 from ..permissions import permission_workflow_edit
 from ..workflow_actions import (
     DocumentPropertiesEditAction, DocumentWorkflowLaunchAction, HTTPAction
@@ -336,7 +336,7 @@ class DocumentWorkflowLaunchActionTestCase(
         action.execute(context={'document': self.test_document})
 
         self.assertTrue(
-            workflow_count + 1, self.test_document.workflows.count()
+            self.test_document.workflows.count(), workflow_count + 1
         )
 
 
@@ -347,6 +347,12 @@ class DocumentWorkflowLaunchActionViewTestCase(
     auto_upload_test_document = False
 
     def test_document_workflow_launch_action_view_with_full_access(self):
+        self._create_test_workflow(add_document_type=True, auto_launch=False)
+        self.test_sub_workflow = self.test_workflow
+        self.test_workflow.internal_name = 'test_sub_workflow'
+        self.test_workflow.label = 'test_sub_workflow'
+        self.test_workflow.save()
+
         self._create_test_workflow(add_document_type=True, auto_launch=False)
         self._create_test_workflow_state()
 
@@ -361,4 +367,34 @@ class DocumentWorkflowLaunchActionViewTestCase(
 
         self.assertEqual(
             self.test_workflow_state.actions.count(), action_count + 1
+        )
+
+    def test_document_workflow_launch_action_view_and_document_create_with_full_access(self):
+        self._create_test_workflow(add_document_type=True, auto_launch=False)
+        self.test_sub_workflow = self.test_workflow
+        self.test_workflow.internal_name = 'test_sub_workflow'
+        self.test_workflow.label = 'test_sub_workflow'
+        self.test_workflow.save()
+
+        self._create_test_workflow(add_document_type=True, auto_launch=True)
+        self._create_test_workflow_state()
+
+        self.grant_access(
+            obj=self.test_workflow, permission=permission_workflow_edit
+        )
+
+        action_count = self.test_workflow_state.actions.count()
+        workflow_instance_count = WorkflowInstance.objects.count()
+
+        response = self._request_document_workflow_launch_action_create_view()
+        self.assertEqual(response.status_code, 302)
+
+        self.assertEqual(
+            self.test_workflow_state.actions.count(), action_count + 1
+        )
+
+        self._create_test_document_stub()
+
+        self.assertEqual(
+            WorkflowInstance.objects.count(), workflow_instance_count + 2
         )